+++ frontend/src/App.vue
@@ -1,9 +1,0 @@
-<script setup>
-import AppLayout from './components/AppLayout.vue';
-import { RouterView } from 'vue-router';
-</script>
-<template>
-  <AppLayout>
-    <RouterView />
-  </AppLayout>
-</template>